Added shared interaction and editor infrastructure
- Added shared interaction, development, database editor, and localization definitions
- Added pen input and pointer-source tracking
- Improved Lua event dispatch
- Fixed Vulkan texture ownership and reference handling

+++ Source/Gameplay/Gameplay/GameDefine.h
@@ -175,7 +175,7 @@ namespace GameDefine
             std::string internalName;
             std::string name;
 
-            u16 type;
+            u8 type;
             u16 maxPlayers;
 
         public:
